Initial version
For the most part this is a copy of files from openstf/stf;
specifically this commit: https://github.com/openstf/stf/commit/d1dd37b320acf11c9d2a11db54faa3203dc9dd5a

Will correctly redo this repo to show the changes, but for now just posted the modified files from the start.

The changes that have been made:
* Update to using all new node modules that are up to date ( copied and pasted mostly from stf-ios-provider for now; there are node modules that aren't needed in there as a result )
* Run on node12 now
* Use built in node events module instead of eventemitter3; required changing all the references
* Use new zeromq in backward compatibility mode
* Use julusian/jpeg-turbo since jpeg-turbo is way out of date and doesn't appear it will ever be updated

Still using out of date protobuf as the way it is setup is now async only, and it wasn't easy to update.
The last version of protobuf 5 also glitches with this set of code; so it has to be manually patched to function properly...
Have not yet figured out the exact cause of it within this code.